requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
requests>=2.31.0
scikit-learn>=1.3.0
plotly>=5.18.0
